Development

+++ Controllers/menuItemController.js
@@ -3,20 +3,22 @@ const MenuItem = require('../db').import('../Models/menuItem');
 const validateToken = require('../Middleware/validateToken');
 
 // Create Menu Item
-router.post('/create', validateToken, (req, res) => {
+router.post('/create', validateToken, async(req, res) => {
     if(req.user.isManager == true){
         const menuItemModel = {
             name: req.body.name,
             price: req.body.price
         };
-
-        MenuItem.create(menuItemModel)
-            .then(item => res.status(200).json({
-                Message: 'Menu Item successfully created',
-
+        try{
+            const item = await MenuItem.create(menuItemModel);
+            res.status(200).json({
+                Message: 'Menu Item Successfully Created',
                 Menu_Item: item
-            }))
-            .catch(err => res.status(500).json({Error: err}));
+            });
+        }
+        catch(err){
+            res.status(500).json({Error: err})
+        }
     }
     else{
         res.status(403).json({Error: 'Not Authorized'});
@@ -24,40 +26,57 @@ router.post('/create', validateToken, (req, res) => {
 });
 
 // Get All Menu Items
-router.get('/all', validateToken, (req, res) => {
-    if(req.user.isManager == true){
-        MenuItem.findAll()
-            .then(items => res.status(200).json({Menu_Items: items}))
-            .catch(err => res.status(500).json({Error: err}));
-    }
-    else{
-        res.status(403).json({Error: 'Not Authorized'});
-    }
+router.get('/all', validateToken, async(req, res) => {
+        try{
+            const items = await MenuItem.findAll({attributes: {exclude: ['createdAt', 'updatedAt']}});
+            if(items.length > 0){
+                res.status(200).json({Menu_Items: items});
+            }
+            else{
+                res.status(200).json({Message: 'No Menu Items Created In System Yet'});
+            }
+        }
+        catch(err){
+            res.status(500).json({Error: err});
+        }
 });
 
 // Get Menu Item By Id
-router.get('/:itemId', validateToken, (req, res) => {
-    if(req.user.isManager == true){
-        MenuItem.findOne({where: {id: req.params.itemId}})
-            .then(item => res.status(200).json({Menu_Item: item}))
-            .catch(err => res.status(500).json({Error: err}));
-    }
-    else{
-        res.status(403).json({Error: 'Not Authorized'});
-    }
+router.get('/:itemId', validateToken, async(req, res) => {
+        try{
+            const item = await MenuItem.findOne({where: {id: req.params.itemId}});
+            if(item){
+                res.status(200).json({Menu_Item: item});
+            }
+            else{
+                res.status(502).json({Error: `No Menu Item Found Matching Item Id: ${req.params.itemId}`});
+            }
+        }
+        catch(err){
+            res.status(500).json({Error: err});
+        }
 });
 
 // Update Menu Item By Id
-router.put('/:itemId', validateToken, (req, res) => {
+router.put('/:itemId', validateToken, async(req, res) => {
     if(req.user.isManager == true){
         const menuItemModel = {
             name: req.body.name,
             price: req.body.price
         };
 
-        MenuItem.update(menuItemModel, {where: {id: req.params.itemId}})
-            .then(() => res.status(200).json({Message: 'Menu Item successfully updated'}))
-            .catch(err => res.status(500).json({Error: err}));
+        if(await MenuItem.findOne({where: {id: req.params.itemId}})){
+            try{
+                await MenuItem.update(menuItemModel, {where: {id: req.params.itemId}});
+                res.status(200).json({Message: 'Menu Item successfully updated'});
+            }
+            catch(err){
+                res.status(500).json({Error: err});
+            }
+        }
+        else{
+            res.status(502).json({Error: `No Menu Item Found Matching Menu Item: ${req.params.itemId}`});
+        }
     }
     else{
         res.status(403).json({Error: 'Not Authorized'});
@@ -65,11 +84,20 @@ router.put('/:itemId', validateToken, (req, res) => {
 });
 
 // Delete Menu Item By Id
-router.delete('/:itemId', validateToken, (req, res) => {
+router.delete('/:itemId', validateToken, async(req, res) => {
     if(req.user.isManager == true){
-        MenuItem.destroy({where: {id: req.params.itemId}})
-            .then(() => res.status(200).json({Message: 'Menu Item successfully removed from system'}))
-            .catch(err => res.status(500).json({Error: err}));
+        try{
+            if(await MenuItem.findOne({where: {id: req.params.itemId}})){
+                await MenuItem.destroy({where: {id: req.params.itemId}});
+                res.status(200).json({Message: 'Menu Item successfully removed from system'});
+            }
+            else{
+                res.status(502).json({Error: `No Menu Item Found Matching Menu Item: ${req.params.itemId}`});
+            }
+        }
+        catch(err){
+            res.status(500).json({Error: err});
+        }
     }
     else{
         res.status(403).json({Error: 'Not Authorized'});
